Sammal v0.1.0 Sammal.ParserCombinators

Basic parser combinator building blocks.

Use these to compose your own grammar.

Types

error()

error() :: atom

parser()

parser() :: (input :: [token] -> {:ok, result} | {:error, remaining :: [token]})

result()

result() :: {value :: [token], remaining :: [token]}

token()

token() :: any

transformer()

transformer() :: ([token] -> [token])

Functions

any(list, errors)

between(a, b, c)

between(parser, parser, parser) :: parser

Parse something between two parsers.

The surrounding parser values are omitted from successful parse result.

both(a, b)

both(parser, parser) :: parser

delay(parser_func)

delay((() -> parser)) :: parser

Wrap parser in a thunk to delay evaluation.

Has the effect of preventing infinite loops in self-referential parsers.

describe(parser, expected)

describe(parser, any) :: parser

Override parser’s expected value that is shown in parse errors.

Examples

iex> parser = Sammal.ParserCombinators.string("x")
iex> parser.(["y"])
{:error, ["y"], "x"}
iex> Sammal.ParserCombinators.describe(parser, "something else").(["y"])
{:error, ["y"], "something else"}

either(a, b)

either(parser, parser) :: parser

eof()

eof() :: parser

Parse end of file.

many(parser)

many(parser) :: parser

many1(parser)

many1(parser) :: parser

number()

number() :: parser

required(parser)

required(parser) :: parser

Turn parser failure into a fatal exception.

sequence(parsers)

sequence([parser]) :: parser

skip(parser)

skip(parser) :: parser

string(string)

string(String.t) :: parser

Parse given string.

transform(f, parser)

transform(transformer, parser) :: parser

Transform successful parse result using given function.

until(parser, stop)

until(parser, parser) :: parser

wrap(parser)

wrap(parser) :: parser

Wrap parser result into a list.